fix(i18n): one language list, shared by the front door and settings
The choice had two designs. Settings showed a script badge per language in its
own brand tint, the chosen row lifted onto the brand wash, and a radio mark on
every row. The public `/language` route — the one reached from the globe on the
welcome and sign-in headers — showed a plain divided list with a bare checkmark
and no badges. That is backwards: the front door is where somebody who cannot
read the current language actually lands, so it is the screen that most needs
colour and script shape to be scannable, and it was the thinner of the two.

The list is now a single `LanguageChoiceList` component holding the restart
banner, the radio group and the footnote. Both screens render it and keep only
what is genuinely theirs: settings keeps the tab-bar clearance and a back that
assumes a stack; the public route keeps the safe-area inset, the brand-ink
chevron, and a back that falls back to `/welcome` when there is no stack.

Also deletes `LanguagePicker`, a compact globe-and-menu third variant of the
same choice that nothing has imported.
